pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "celestialflow"                   
version = "3.1.6"
description = "A flexible GRAPH-based task orchestration framework."
readme = "README.md"               
license = { text = "MIT" }    
authors = [{ name = "Mr-xiaotian", email = "mingxiaomingtian@gmail.com" }]
keywords = ["workflow", "task", "graph", "async", "CelestialFlow"]
requires-python = ">=3.10"

dependencies = [
    "tqdm",
    "fastapi",
    "uvicorn",
    "requests",
    "networkx",
    "redis",
    "jinja2",
    "celestialtree>=0.1.2",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries",
]

[project.urls]
"Homepage" = "https://github.com/Mr-xiaotian/CelestialFlow"
"Bug Tracker" = "https://github.com/Mr-xiaotian/CelestialFlow/issues"

[project.scripts]
celestialflow-web = "celestialflow.web.server:main_entry"

[tool.setuptools]
license-files = []

[tool.setuptools.package-data]
celestialflow = [
    "web/templates/**/*.html",
    "web/static/**/*",
    "web/config.json"
]

[tool.pytest.ini_options]
addopts = "-s"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestDeprecationWarning",
]

asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "twine>=6.1.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "black>=26.1.0",
]
